+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ### Bugfixes
 * (De)serializing of blank node Entities with Constrains did not keep the proper link
+* Allow to change display order of Variable components
 
 ## 0.3.0